Move issue_178 plugin to conftest.py to avoid bootstrap errors on Win
